requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-storage
python-multipart
